requirements.txt
snowflake-connector-python>=2.7.0
PyYAML>=6.0
pandas>=1.3.0
python-dotenv>=0.19.0